Extract text watcher class.
